# openAPI.yaml
openapi: 3.0.3
info:
  title: Todo List App
  description: OPEN API for Todo tasks API
  version: 1.0.0
  contact:
    name: Mahmuda Karima (DAKA)
    email: dakasakti.id@gmail.com
  license:
    name: APACHE 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  # Added by API Auto Mocking Plugin
  - description: Todo tasks API Server
    url: https://{environment}.todoapps.com/api/v1/
    variables:
      environment:
        default: dev
        enum:
          - dev
          - prod
paths:
...
  /tasks/{id}/completed:
    post:
      security:
        - JWTAuth: []
      parameters:
        - $ref: "#/components/parameters/TodolistId"
      tags:
        - Todo tasks
      summary: Reopen todo tasks
      description: Reopen todo list task
      responses:
        "200":
          $ref: "#/components/responses/Response200"
        "400":
          $ref: "#/components/responses/Response400"
...
components:
  securitySchemes:
    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TodolistId:
      name: id
      in: path
      required: true
      description: Todolist id for update
      schema:
        type: integer
...
  responses:
    Response200:
      description: success create
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: number
                example: 200
              message:
                type: string
                example: successfully
              data:
                type: object
                example: {}
...
    Response400:
      description: bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: number
                example: 400
              message:
                type: string
                example: bad request (bind, validate, duplicate)
              data:
                type: object
                example: null
...